+++ util/miscellaneousConsts.js
@@ -1,0 +1,1 @@
+export const CURRENT_SCHEMA = "1.0.0";